--- purchase_order_date/views/order_date.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record model="ir.ui.view" id="order_date_form">
-            <field name="name">order.date.form</field>
-            <field name="model">purchase.order</field>
-            <field name="inherit_id" ref="purchase.purchase_order_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='date_order']" position="after">
-                    <!-- <field name="date_orders"/> -->
-                    <field name="date_orders_old" invisible="1"/>
-                </xpath>
-                <!-- <xpath expr="//field[@name='company_id']" position="after">
-                    <field name="qty_rec" attrs="{'invisible':[('state', 'not in', ['purchase'])]}"/>
-                </xpath> -->
-            </field>
-        </record>
-    </data>
-</odoo>
